renew: 전면 롤백
