--- src/main/java/dapanda/adapter/common/NotFoundException.java
@@ -1,14 +1,0 @@
-package dapanda.adapter.common;
-
-import lombok.Getter;
-
-@Getter
-public class NotFoundException extends NullPointerException {
-
-    private final int code;
-
-    public NotFoundException(final ErrorType errorType) {
-        super(errorType.getMessage());
-        this.code = errorType.getCode();
-    }
-}
