--- src/main/java/dapanda/adapter/inbound/controller/request/WebRequest.java
@@ -1,7 +1,0 @@
-package dapanda.adapter.inbound.controller.request;
-
-public class WebRequest {
-
-        public record OrderProductWebRequestV1(int orderAmount, String categoryType) {
-    }
-}
