+++ src/main/java/dapanda/application/store/StoreService.java
@@ -1,12 +1,11 @@
 package dapanda.application.store;
 
-import dapanda.adapter.common.NotFoundException;
+import dapanda.domain.common.NotFoundException;
+import dapanda.domain.outbound.jpa.order.OrderEntity;
+import dapanda.domain.outbound.jpa.order.repository.OrderJpaRepository;
+import dapanda.domain.outbound.jpa.store.StoreEntity;
+import dapanda.domain.outbound.jpa.store.repository.StoreJpaRepository;
 import dapanda.application.store.dto.StoreServiceDto;
-import dapanda.domain.order.Order;
-import dapanda.domain.store.Store;
-import dapanda.domain.store.StoreRepository;
-import dapanda.domain.store.product.Product;
-import dapanda.domain.store.product.ProductRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -14,46 +13,40 @@ import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 
-import static dapanda.adapter.common.ErrorType.NOT_FOUND_PRODUCT_INFO;
-import static dapanda.adapter.common.ErrorType.NOT_FOUND_STORE_INFO;
+import static dapanda.domain.common.ErrorType.NOT_FOUND_STORE_INFO;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
 public class StoreService {
 
-    private final StoreRepository storeRepository;
-    private final ProductRepository productRepository;
+    private final StoreJpaRepository storeRepository;
+    private final OrderJpaRepository orderRepository;
 
     @Transactional
     public void order(final StoreServiceDto.OrderDto dto) {
 
-        Optional<Store> store = getStoreInfoById(dto.storeId());
-        Optional<Product> product = getProductInfoById(dto.productId());
-
-
+        StoreEntity store = getStoreInfoById(dto.storeId());
     }
 
-    private Store getStoreInfoById(final long storeId) {
+    private StoreEntity getStoreInfoById(final long storeId) {
         return storeRepository.findById(storeId)
                 .orElseThrow(() -> new NotFoundException(NOT_FOUND_STORE_INFO));
     }
 
-    private Optional<Product> getProductInfoById(final long productId) {
-        Optional<Product> product = productRepository.findById(productId);
-
-        if (product == null) {
-            throw new NotFoundException(NOT_FOUND_PRODUCT_INFO);
-        }
-        return product;
-    }
-
     @Transactional(readOnly = true)
-    public Order findOrder(final long storeId, final long orderId) {
+    public Optional<OrderEntity> findOrder(final long storeId, final long orderId) {
 
         log.info("{} {}", storeId, orderId);
 
-        Store store = getStoreInfoById(storeId);
-        return store.getOrderById(orderId);
+        return orderRepository.findById(orderId);
+    }
+
+    @Transactional(readOnly = true)
+    public Optional<StoreEntity> findStore(final long storeId) {
+
+        log.info("{} {}", storeId);
+
+        return storeRepository.findById(storeId);
     }
 }
