+++ src/main/java/dapanda/domain/common/InvalidInputException.java
@@ -1,0 +1,14 @@
+package dapanda.domain.common;
+
+import lombok.Getter;
+
+@Getter
+public class InvalidInputException extends RuntimeException {
+
+    private final int code;
+
+    public InvalidInputException(final ErrorType errorType) {
+        super(errorType.getMessage());
+        this.code = errorType.getCode();
+    }
+}
