+++ src/main/java/dapanda/domain/common/NotFoundException.java
@@ -1,0 +1,14 @@
+package dapanda.domain.common;
+
+import lombok.Getter;
+
+@Getter
+public class NotFoundException extends NullPointerException {
+
+    private final int code;
+
+    public NotFoundException(final ErrorType errorType) {
+        super(errorType.getMessage());
+        this.code = errorType.getCode();
+    }
+}
