--- src/main/java/dapanda/domain/customer/CustomerRepository.java
@@ -1,10 +1,0 @@
-package dapanda.domain.customer;
-
-import java.util.Optional;
-
-public interface CustomerRepository {
-
-    public Customer save(Customer customer);
-
-    public Optional<Customer> findById(long id);
-}
