+++ src/main/java/dapanda/domain/inbound/controller/request/WebRequest.java
@@ -1,0 +1,7 @@
+package dapanda.domain.inbound.controller.request;
+
+public class WebRequest {
+
+        public record OrderProductWebRequestV1(int orderAmount, String categoryType) {
+    }
+}
