+++ src/main/java/dapanda/domain/inbound/controller/response/WebResponse.java
@@ -1,0 +1,4 @@
+package dapanda.domain.inbound.controller.response;
+
+public class WebResponse {
+}
