--- src/main/java/dapanda/domain/order/OrderRepository.java
@@ -1,10 +1,0 @@
-package dapanda.domain.order;
-
-import java.util.Optional;
-
-public interface OrderRepository {
-
-    public Order save(Order order);
-
-    public Optional<Order> findById(long id);
-}
