+++ src/main/java/dapanda/domain/outbound/jpa/BaseEntity.java
@@ -1,0 +1,33 @@
+package dapanda.domain.outbound.jpa;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@EntityListeners(AuditingEntityListener.class)
+@MappedSuperclass
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+public abstract class BaseEntity {
+
+    @Column(nullable = false, updatable = false, name = "created_at")
+    @CreatedDate
+    private LocalDateTime createdAt;
+
+    @Column(nullable = false, name = "last_modified")
+    @LastModifiedDate
+    private LocalDateTime lastModified;
+
+    protected BaseEntity(final LocalDateTime createdAt, final LocalDateTime lastModified) {
+        this.createdAt = createdAt;
+        this.lastModified = lastModified;
+    }
+}
