+++ src/main/java/dapanda/domain/outbound/jpa/customer/CustomerEntity.java
@@ -1,0 +1,54 @@
+package dapanda.domain.outbound.jpa.customer;
+
+import dapanda.domain.outbound.jpa.BaseEntity;
+import dapanda.domain.outbound.jpa.store.StoreEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Entity
+@Table(name = "customer")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class CustomerEntity extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+    private String name;
+    private String email;
+    private String password;
+    private String tel;
+
+    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.REFRESH)
+    private StoreEntity store;
+
+    private CustomerEntity(
+            final long id,
+            final String name,
+            final String email,
+            final String password,
+            final String tel
+    ) {
+        super(LocalDateTime.now(), LocalDateTime.now());
+
+        this.id = id;
+        this.name = name;
+        this.email = email;
+        this.password = password;
+        this.tel = tel;
+    }
+
+    public static CustomerEntity of(
+            final long id,
+            final String name,
+            final String email,
+            final String password,
+            final String tel
+    ) {
+        return new CustomerEntity(id, name, email, password, tel);
+    }
+}
