+++ src/main/java/dapanda/domain/outbound/jpa/customer/repository/CustomerJpaRepository.java
@@ -1,0 +1,7 @@
+package dapanda.domain.outbound.jpa.customer.repository;
+
+import dapanda.domain.outbound.jpa.customer.CustomerEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CustomerJpaRepository extends JpaRepository<CustomerEntity, Long> {
+}
