+++ src/main/java/dapanda/domain/outbound/jpa/store/product/repository/ProductJpaRepository.java
@@ -1,0 +1,7 @@
+package dapanda.domain.outbound.jpa.store.product.repository;
+
+import dapanda.domain.outbound.jpa.store.product.ProductEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductJpaRepository extends JpaRepository<ProductEntity, Long> {
+}
