--- src/main/java/dapanda/domain/store/StoreCategoryType.java
@@ -1,15 +1,0 @@
-package dapanda.domain.store;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum StoreCategoryType {
-
-    FOOD("food"),
-    CLOTH("cloth")
-    ;
-
-    private final String typeName;
-}
