--- src/main/java/dapanda/domain/store/product/ProductRepository.java
@@ -1,10 +1,0 @@
-package dapanda.domain.store.product;
-
-import java.util.Optional;
-
-public interface ProductRepository {
-
-    public Product save(Product product);
-
-    public Optional<Product> findById(long id);
-}
